requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
cloudscraper==1.2.58
idna==3.3
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyparsing==2.4.7
requests==2.26.0
requests-toolbelt==0.9.1
toml==0.10.2
urllib3==1.26.7
